move TrackNavigationMessage implementation
