+++ Screenbox.Core/ViewModels/PlayerControlsViewModel.cs
@@ -24,6 +24,7 @@ public sealed partial class PlayerControlsViewModel : ObservableRecipient,
     IRecipient<PropertyChangedMessage<IMediaPlayer?>>,
     IRecipient<SettingsChangedMessage>,
     IRecipient<TogglePlayPauseMessage>,
+    IRecipient<TrackNavigationMessage>,
     IRecipient<ChangePlaybackRateRequestMessage>,
     IRecipient<PropertyChangedMessage<PlayerVisibilityState>>,
     IRecipient<PropertyChangedMessage<WindowViewMode>>
@@ -110,6 +111,7 @@ public sealed partial class PlayerControlsViewModel : ObservableRecipient,
         Messenger.Register<PropertyChangedMessage<IMediaPlayer?>>(this);
         Messenger.Register<SettingsChangedMessage>(this);
         Messenger.Register<TogglePlayPauseMessage>(this);
+        Messenger.Register<TrackNavigationMessage>(this);
         Messenger.Register<ChangePlaybackRateRequestMessage>(this);
         Messenger.Register<PropertyChangedMessage<PlayerVisibilityState>>(this);
         Messenger.Register<PropertyChangedMessage<WindowViewMode>>(this);
@@ -154,6 +156,30 @@ public sealed partial class PlayerControlsViewModel : ObservableRecipient,
         else
         {
             PlayPause();
+        }
+    }
+
+    public async void Receive(TrackNavigationMessage message)
+    {
+        if (MediaPlayer is null)
+            return;
+
+        switch (message.Direction)
+        {
+            case TrackNavigationDirection.Next:
+                if (CanGoNext())
+                {
+                    await _coordinator.NextAsync();
+                }
+
+                break;
+            case TrackNavigationDirection.Previous:
+                if (CanGoPrevious())
+                {
+                    await _coordinator.PreviousAsync();
+                }
+
+                break;
         }
     }
